commands: Move atomic unit refusals
Batch-source action modules already own command-facing refusal helpers for
merge candidates and worktree conflicts. Batch selection still formats
atomic ownership errors with session display ID ranges.

The project does not provide a command-owned place for atomic ownership
refusals. Keeping that formatter in batch.selection makes the batch
selection module depend on session progress display formatting.

This commit addresses that by moving atomic ownership refusal formatting
into commands.batch_source.atomic_unit_refusals and updating apply,
include, and discard batch-source actions to use that support module.

The final commit will pin that command-facing ownership in architecture
tests.
